not even close

+++ src/front/odGUI/OverdoseDeathsTable.svelte
@@ -52,13 +52,26 @@
 		
 	}
 
+	async function deleteOverdoseDeaths() {
+		if(confirm("¿Está seguro de que desea eliminar todas las entradas?")){
+			console.log("Deleting all overdose deaths...");
+			const res = await fetch("/api/v1/overdose-deaths/", {
+				method: "DELETE"
+			}).then(function (res) {
+				getOverdoseDeaths();
+			});
+		}
+	}
+
 	async function deleteOverdoseDeath(country,year) {
-		console.log("Deleting overdose death...");
-		const res = await fetch("/api/v1/overdose-deaths/" + country + "/"+year,{
-			method:"DELETE"
-		}).then(function (res) {
-			getOverdoseDeaths();
-		});	
+		if(confirm("¿Está seguro de que desea eliminar esta entrada?")){
+			console.log("Deleting overdose death...");
+			const res = await fetch("/api/v1/overdose-deaths/" + country + "/"+year,{
+				method:"DELETE"
+			}).then(function (res) {
+				getOverdoseDeaths();
+			});	
+		}
 	}
 	</script>
 
@@ -74,7 +87,7 @@
 					<th>Año</th>
 					<th>Hombres fallecidos</th>
 					<th>Mujeres fallecidas</th>
-					<th>Total de fallecidos</th>
+					<th>Total fallecidos</th>
 					<th>Edad media</th>
 					<th>Actions</th>
 				</tr>
@@ -86,8 +99,8 @@
 						<td><input type="number" placeholder="20" min=0 bind:value="{newOverdoseDeath.death_male}"></td>
 						<td><input type="number" placeholder="10" min=0 bind:value="{newOverdoseDeath.death_female}"></td>
 						<td><input type="number" placeholder="30" min=0 bind:value="{newOverdoseDeath.death_total}"></td>
-						<td><input type="number" placeholder="20.5" step=0.1 min=0 bind:value="{newOverdoseDeath.mean_age}"></td>
-						<td><Button outline color= "primary"  on:click={insertOverdoseDeath}>Insert</Button></td>
+						<td><input type="number" placeholder="20.5" min=0 step=0.1  bind:value="{newOverdoseDeath.mean_age}"></td>
+						<td><Button outline color= "primary"  on:click={insertOverdoseDeath}>Insertar</Button></td>
 					</tr>
 				{#each overdose_deaths as overdose_death}
 					<tr>
@@ -99,10 +112,11 @@
 						<td>{overdose_death.death_female}</td>
 						<td>{overdose_death.death_total}</td>
 						<td>{overdose_death.mean_age}</td>
-						<td><Button outline color="danger" on:click="{deleteOverdoseDeath(overdose_death.country, overdose_death.year)}">Delete</Button></td>
+						<td><Button outline color="danger" on:click="{deleteOverdoseDeath(overdose_death.country, overdose_death.year)}">Eliminar</Button></td>
 					</tr>
 				{/each}
 			</tbody>
 		</Table>
 	{/await}
+	<Button outline on:click={deleteOverdoseDeaths} color="danger">  Eliminar todo </Button>
 </main>
